// FilmLibraryTest.Web/ClientApp/src/apiClient/apiClient.ts
export interface IResultModelOfListFilm {
    error?: string | undefined;
    result?: ListFilm | undefined;
}

export interface ListFilm {
    films?: FilmViewModel[] | undefined;
    totalCount: number;
}

export interface FilmViewModel {
    id: number;
    title?: string | undefined;
    description?: string | undefined;
    year: number;
    producer?: string | undefined;
    userName?: string | undefined;
}